This version combines the two scripts that
- extract out the books and related methods
- use `book_folder/1` to build links

I'm not sure how I feel about this... this means I need to do something
fancy on the Drafts/writing side in order to get the book numbers in
order. I don't think that's helpful for my goal of "fast recording." I
think I'm going to remove the ordering numbers on the books, and merely
have an index on The Bible note.